Attempting to pass Lat and Lon values to updated Calculator class

+++ INB372-Solar/src/unitTests/DistanceCalcTests.java
@@ -13,20 +13,20 @@ public class DistanceCalcTests {
 	@Test
 	public void normalCloseBrisban() {
 		DistanceCalc brisClose = new DistanceCalc();
-		float morayfieldLat = -27.105453f;
-		float morayfieldLon = 152.948145f;
-		float[] closestSunTable = brisClose.findClosestStation(morayfieldLat, morayfieldLon);
-		assert(closestSunTable[0] == 3.3f);
+		double morayfieldLat = -27.105453;
+		double morayfieldLon = 152.948145;
+		double[] closestSunTable = brisClose.findClosestStation(morayfieldLat, morayfieldLon);
+		assert(closestSunTable[0] == 3.3);
 		
 	}
 	
 	@Test
 	public void normalClosePerth() {
 		DistanceCalc brisClose = new DistanceCalc();
-		float northamLat = -31.65453f;
-		float northamLon = 116.670517f;
-		float[] closestSunTable = brisClose.findClosestStation(northamLat, northamLon);
-		assert(closestSunTable[0] == 9.4f);
+		double northamLat = -31.65453;
+		double northamLon = 116.670517;
+		double[] closestSunTable = brisClose.findClosestStation(northamLat, northamLon);
+		assert(closestSunTable[0] == 9.4);
 	}
 
 }
